Move errors to a separate module
